delete

+++ src/main/java/com/Hansung/Capston/service/Diet/IngredientService.java
@@ -85,10 +85,14 @@ public class IngredientService {
   }
 
   // 식재료 즐겨찾기 삭제하기
-  @Transactional
+  @Transactionalg
   public String deletePreferredIngredient(Long preferredIngredientId) {
-    recommendedIngredientRepository.deleteById(preferredIngredientId);
-    return("성공 : 즐겨찾기 삭제");
+    if (recommendedIngredientRepository.existsById(preferredIngredientId)) {
+      recommendedIngredientRepository.deleteById(preferredIngredientId);
+      return "성공 : 즐겨찾기 삭제";
+    } else {
+      return "실패 : 해당 즐겨찾기 ID가 존재하지 않습니다.";
+    }
   }
 
   // 추천 식재료 등록하기
